requirements.txt
Flask==1.0.2
Bootstrap_Flask==1.5.1
pygments==2.7.2
python-dotenv==0.15.0
selenium==3.141.0